Set words per minute on page to the calculated number and round it
The calculated number of words typed per minute is set as the value of the words per minute displayed on the page. A function is used to round this value to an integer because a result of 40.464 words per minute seems odd and unpolished. The error message is set back to nothing if the user had already submitted and received an error message.